UPDATE -Renamed MainHome to Pricing
