+++ src/pages/Pricing.tsx
